requirements.txt
openai-whisper
torch
ffmpeg-python
imageio-ffmpeg
transformers
sentencepiece
